// src/lib/db/type.ts
export interface Users {
  users: string;
  addNewUser: string;
  noResults: string;
  NAME: string;
  ROLE: string;
  BRANCH: string;
  ACTIVITY: string;
  NoBrowser: string;
  ACTION: string;
  LastLogin: string;
  Browser: string;
  Search: string;
  name: string;
  anUnexpectedErrorOccurred: string;
  Success: string;
  Error: string;
  Avatar: string;
  FirstName: string;
  LastName: string;
  Email: string;
  EmailAddress: string;
  Role: string;
  SelectARole: string;
  Agent: string;
  Manager: string;
  Director: string;
  Branch: string;
  SelectABranch: string;
  NoBranchesFound: string;
  SearchForABranch: string;
  PhoneNumber: string;
  messenger: string;
  viber: string;
  whatsApp: string;
  telegram: string;
  AddNewNumber: string;
  Bio: string;
  description: string;
  AddressAndLocation: string;
  AddLocationAndAddress: string;
  StreetAddress: string;
  Country: string;
  City: string;
  ZipCode: string;
  State: string;
  SaveChanges: string;
  Create: string;
  EditUser: string;
  AddUser: string;
  EditAUser: string;
  AddANewUser: string;
  firstNameError: string;
  lastNameError: string;
  emailError: string;
  bioError: string;
  branchIdError: string;
  streetError: string;
  cityError: string;
  stateError: string;
  zipCodeError: string;
  countryError: string;
  phoneNumberError: string;
  phoneTypeError: string;
  roleError: string;
  FailedToLoadUserDetails: string;
  NoPropertiesFoundForThisUser: string;
  Back: string;
  steet: string;
  AgentDetails: string;
  IdentificationNumber: string;
  PleaseWait: string;
  Deleting: string;
  UserSuccessfullyDeleted: string;
  SomethingWentWrongPleaseTryAgainLater: string;
  ImageSizeLimit: string;
  ImageExceedsSizeLimit: string;
  DropOrUploadPhoto: string;
}
export interface BranchTypes {
  Branches: string;
  BRANCH: string;
  PHONENUMBER: string;
  ADDRESS: string;
  ACTION: string;
  Search: string;
  name: string;
  AddNewBranch: string;
  NoResults: string;
  EditBranch: string;
  SaveChanges: string;
  CreateBranch: string;
  error: string;
  success: string;
  BackToBranch: string;
  CoverImage: string;
  Name: string;
  BranchName: string;
  Email: string;
  EmailAddress: string;
  Phone: string;
  PhoneNumber: string;
  AddressAndLocation: string;
  AddLocationandaddress: string;
  StreetAddress: string;
  City: string;
  State: string;
  ZipCode: string;
  Country: string;
  branchNameError: string;
  coverImageError: string;
  streetError: string;
  cityError: string;
  stateError: string;
  zipCodeError: string;
  countryError: string;
  phoneError: string;
  emailError: string;
  faxError: string;
  websiteError: string;
  Back: string;
  BranchDetails: string;
  IdentificationNumber: string;
  PleaseWait: string;
  Deleting: string;
  BranchSuccessfullyDeleted: string;
  SomethingWentWrongPleaseTryAgainLater: string;
  ImageSizeLimit: string;
  ImageExceedsSizeLimit: string;
  DropOrUploadPhoto: string;
}